+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Remember to align the itemized text with the first line of an item within a list
 * Deprecations
   * The `in_axis_resources` and `out_axis_resources` arguments of pjit have been
     deprecated. Please use `in_shardings` and `out_shardings` respectively.
+  * The function `jax.numpy.msort` has been removed. It has been deprecated since
+    JAX v0.4.1. Use `jnp.sort(a, axis=0)` instead.
 
 ## jaxlib 0.4.9